refactor: IssuedLotto의 컨트롤러 의존성 변경
